Simplify the split edit code (#1904)
* Simplify the split edit code

* span -> div
